requirements.txt
# Core dependencies
requests
beautifulsoup4
pandas
fastapi
uvicorn
python-dotenv
google-cloud-bigquery
pyarrow
anthropic
apache-airflow[gcp]==2.9.1
pendulum
prometheus_client

# Test dependencies (consolidées depuis tests/requirements.txt)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
great-expectations>=0.17.0
pydantic>=2.0.0
responses>=0.23.0
httpx>=0.24.0
psutil>=5.9.0
pytest-benchmark>=4.0.0
pylint>=2.17.0
black>=23.0.0
mypy>=1.5.0
